Apply react-native-modal, Modify GroupCreateImage
