+++ package.json
@@ -25,6 +25,7 @@
     "react-native-gesture-handler": "1.10.3",
     "react-native-image-crop-picker": "^0.37.2",
     "react-native-linear-gradient": "2.5.6",
+    "react-native-modal": "^13.0.0",
     "react-native-reanimated": "2.2.4",
     "react-native-safe-area-context": "3.3.2",
     "react-native-screens": "3.9.0",
