+++ src/containers/group/create/components/TextInputGroup.js
@@ -18,7 +18,8 @@ const styles = StyleSheet.create({
     justifyContent : 'center',
     width : '100%',
     height : 240,
-    paddingHorizontal : 18,
+    paddingHorizontal: 18,
+    marginBottom: 146,
   }
 })
